[refactor] user 엔티티 테이블명 수정, repository 패키지 생성
