+++ easytask/src/main/java/com/easytask/easytask/src/user/UserService.java
@@ -2,13 +2,14 @@ package com.easytask.easytask.src.user;
 
 import com.easytask.easytask.common.exception.BaseException;
 import com.easytask.easytask.common.jwt.TokenProvider;
-import com.easytask.easytask.src.user.dto.requestDto.UserLoginDto;
-import com.easytask.easytask.src.user.dto.requestDto.UserRequestDto;
-import com.easytask.easytask.src.user.dto.responseDto.UserResponseDto;
+import com.easytask.easytask.src.user.dto.request.UserLoginDto;
+import com.easytask.easytask.src.user.dto.request.UserRequestDto;
+import com.easytask.easytask.src.user.dto.response.UserResponseDto;
 import com.easytask.easytask.src.user.entity.PossibleTask;
 import com.easytask.easytask.src.user.entity.Role;
 import com.easytask.easytask.src.user.entity.TaskAbility;
 import com.easytask.easytask.src.user.entity.User;
+import com.easytask.easytask.src.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
