+++ easytask/src/main/java/com/easytask/easytask/src/user/dto/request/TaskRequestDto.java
@@ -1,12 +1,10 @@
-package com.easytask.easytask.src.user.dto.requestDto;
+package com.easytask.easytask.src.user.dto.request;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 @Getter
