+++ easytask/src/main/java/com/easytask/easytask/src/user/dto/request/UserRequestDto.java
@@ -1,7 +1,5 @@
-package com.easytask.easytask.src.user.dto.requestDto;
+package com.easytask.easytask.src.user.dto.request;
 
-import com.easytask.easytask.src.user.entity.Role;
-import com.easytask.easytask.src.user.entity.User;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
